+++ CHANGELOG.md
@@ -4,6 +4,12 @@ All notable changes to this project will be documented in this file.
 
 ## [Unreleased]
 
+### Added
+- Add command to pull change log from CHANGELOG.md
+
+### Changes
+- Run publish build when tag provided in format [0-9]*.[0-9]*.[0-9]*
+
 ---
 
 ## [0.0.7] - 2025-04-15